+++ stubs-electropol-fr/lib/microbit.py
@@ -15,29 +15,28 @@ _UARTParity = NewType("_UARTParity", int)
 
 def panic() -> None:
     """
-    i18n.t('microbit.put-micro-bit-in-panic-mode-and-display-an-unhappy-face-press-the-reset-button-to-exit-panic-mode')
+    Met la micro:bit en mode panique et affichez un visage mécontent. Appuyer sur le bouton de réinitialisation pour quitter le mode panique.
     """
     pass
 
 
-def sleep(time: int) -> None:
+def sleep(durée: int) -> None:
     """
-    Put micro:bit to sleep for some milliseconds (1 second = 1000 ms) of time.
-    sleep(2000) gives micro:bit a 2 second nap.
+    Mettre la micro:bit en sommeil pour n millisecondes (1 seconde = 1000 ms). 
     """
     pass
 
 
 def running_time() -> int:
     """
-    Return running_time() in milliseconds since micro:bit's last reset.
+    Renvoie la durée en millisecondes depuis la dernière réinitialisation de micro:bit.
     """
     pass
 
 
 def temperature() -> float:
     """
-    Return micro:bit's temperature in degrees Celcius.
+    Retourne la température en degrés Celsius.
     """
     pass
 
@@ -46,51 +45,50 @@ def temperature() -> float:
 class _Accelerometer(object):
     def get_x(self) -> int:
         """
-        Return micro:bit's tilt (X acceleration) in milli-g's.
+        Retourne l'accélération x en mg.
         """
         pass
 
     def get_y(self) -> int:
         """
-        Return micro:bit's tilt (Y acceleration) in milli-g's.
+        Retourne l'accélération y en mg.
         """
         pass
 
     def get_z(self) -> int:
         """
-        Return micro:bit's up-down motion (Z acceleration) in milli-g's.
-        Z is a positive number when moving up. Moving down, Z is a negative
-        number.
+        Retourne l'accélération z en mg.
+        Z est positif lors d'un mouvement vers le haut et négatif vers le bas
         """
         pass
 
-    def is_gesture(self, name: str) -> bool:
+    def is_gesture(self, nom: str) -> bool:
         """
-        Return True or False to indicate if the named gesture is currently
-        active.
-        MicroPython understands the following gestures: 'up', 'down', 'left',
-        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' and
+        Retourne vrai ou faux pour indiquer si le geste nommé est actuellement 
+        actif. 
+        Micropython comprend les gestes suivants: 'up', 'down', 'left',
+        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' et
         'shake'.
         """
         pass
 
-    def was_gesture(self, name: str) -> bool:
+    def was_gesture(self, nom: str) -> bool:
         """
-        Return True or False to indicate if the named gesture was active since
-        the last call.
-        MicroPython understands the following gestures: 'up', 'down', 'left',
-        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' and
+        Retourne vrai ou faux pour indiquer si le geste nommé était actif depuis 
+        le dernier appel.
+        Micropython comprend les gestes suivants: 'up', 'down', 'left',
+        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' et
         'shake'.
         """
         pass
 
     def get_gestures(self) -> List[str]:
         """
-        Return a list indicating the gesture history. The most recent gesture
-        is last.
-        Calling this method also clears the gesture history.
-        MicroPython understands the following gestures: 'up', 'down', 'left',
-        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' and
+        Renvoie une liste indiquant l'historique des gestes. Le geste le plus
+        récent est le dernier. 
+        Appeler cette méthode efface également l'historique des gestes.
+        Micropython comprend les gestes suivants: 'up', 'down', 'left',
+        'right', 'face up', 'face down', 'freefall', '3g', '6g', '8g' et
         'shake'.
         """
         pass
@@ -103,21 +101,21 @@ accelerometer = _Accelerometer()
 class _Button(object):
     def is_pressed(self) -> bool:
         """
-        If the button is pressed down, is_pressed() is True, else False.
+        Si le bouton est enfoncé, is_pressed() est vrai, sinon faux.
         """
         pass
 
     def was_pressed(self) -> bool:
         """
-        Use was_pressed() to learn if the button was pressed since the last
-        time was_pressed() was called. Returns True or False.
+        Utilisez was_pressed() pour savoir si le bouton a été pressé depuis la dernière
+        fois que was_pressed() a été appelé. Renvoie Vrai ou Faux.
         """
         pass
 
     def get_presses(self) -> int:
         """
-        Use get_presses() to get the running total of button presses, and also
-        reset this counter to zero.
+        Utilisez get_presses() pour obtenir le nombre total de pressions sur les boutons, et aussi
+        remettre ce compteur à zéro.
         """
         pass
 
@@ -130,65 +128,57 @@ button_b = _Button()
 class _Compass(object):
     def is_calibrated(self) -> bool:
         """
-        If micro:bit's compass is_calibrated() and adjusted for accuracy,
-        return True.
-        If compass hasn't been adjusted for accuracy, return False.
+        Si la boussole est calibrée, renvoie Vrai.
+        Sinon, renvoie False.
         """
         pass
 
     def calibrate(self) -> None:
         """
-        If micro:bit is confused, calibrate() the compass to adjust the its
-        accuracy.
-        Will ask you to rotate the device to draw a circle on the display.
-        Afterwards, micro:bit will know which way is north.
+        Calibre la boussole pour ajuster sa précision.
+        Vous demandera de faire pivoter l'appareil pour dessiner un cercle sur l'écran.
+        Ensuite, micro:bit saura où est le nord.
         """
         pass
 
     def clear_calibration(self) -> None:
         """
-        Reset micro:bit's compass using clear_calibration() command.
-        Run calibrate() to improve accuracy.
+        Réinitialisez la boussole du micro:bit à l'aide de la commande clear_calibration().
+        Exécutez calibrate() pour améliorer la précision.
         """
         pass
 
     def get_x(self) -> int:
         """
-        Return magnetic field detected along micro:bit's X axis.
-        Usually, the compass returns the earth's magnetic field in micro-Tesla
-        units.
-        Unless...a strong magnet is nearby!
+        Retourne le champ magnétique détecté sur l'axe X du micro:bit.
+        En microtesla.
         """
         pass
 
     def get_y(self) -> int:
         """
-        Return magnetic field detected along micro:bit's Y axis.
-        Usually, the compass returns the earth's magnetic field in micro-Tesla
-        units.
-        Unless...a strong magnet is nearby!
+        Retourne le champ magnétique détecté sur l'axe Y du micro:bit.
+        En microtesla.
         """
         pass
 
     def get_z(self) -> int:
         """
-        Return magnetic field detected along micro:bit's Z axis.
-        Usually, the compass returns the earth's magnetic field in micro-Tesla
-        units.
-        Unless...a strong magnet is nearby!
+        Retourne le champ magnétique détecté sur l'axe Z du micro:bit.
+        En microtesla.
         """
         pass
 
     def get_field_strength(self) -> int:
         """
-        Return strength of magnetic field around micro:bit.
+        Renvoie la force du champ magnétique autour de micro:bit.
         """
         pass
 
     def heading(self) -> int:
         """
-        Return a number between 0-360 indicating the device's heading. 0 is
-        north.
+        Renvoie un nombre compris entre 0 et 360 indiquant l'orientation de la carte. 
+        0 pour le Nord.
         """
         pass
 
@@ -200,80 +190,77 @@ compass = _Compass()
 class _Display(object):
     def show(
         self,
-        x: Union['Image', List['Image'], str],
-        delay: int = 400,
-        wait: bool = True,
-        loop: bool = False,
-        clear: bool = False,
+        monImage: Union['Image', List['Image'], str],
+        retard: int = 400,
+        attendre: bool = True,
+        boucle: bool = False,
+        efface: bool = False,
     ) -> None:
         """
-        Use show(x) to print the string or image 'x' to the display. If 'x' is
-        a list of images they will be animated together.
-        Use 'delay' to specify the speed of frame changes in milliseconds.
-        If wait is False animation will happen in the background while the
-        program continues.
-        If loop is True the animation will repeat forever.
-        If clear is True the display will clear at the end of the animation.
+        Utilisez Show(monImage) pour afficher la chaîne ou l'image «monImage».
+        Si «monImage» est une liste d'images, elles seront animées ensemble. 
+        Utilisez le « retard » pour spécifier les modifications de la vitesse
+        d'affichage en millisecondes. 
+        Si attendre est faux, l'animation s'exécutera pendant que le programme
+        continue.
+        Si boucle est vrai, l'animation se répétera indéfiniment. 
+        Si efface est vrai, l'écran s'éteindra à la fin de l'animation.
         """
         pass
 
     def scroll(
         self,
-        string: str,
-        delay: int = 150,
-        wait: bool = True,
-        loop: bool = False,
+        maChaine: str,
+        retard: int = 150,
+        attendre: bool = True,
+        boucle: bool = False,
         monospace: bool = False,
     ) -> None:
         """
-        Use scroll(string) to scroll the string across the display.
-        Use delay to control how fast the text scrolls.
-        If wait is False the text will scroll in the background while the
-        program continues.
-        If loop is True the text will repeat forever.
-        If monospace is True the characters will always take up 5
-        pixel-columns.
+        Utilisez Scroll(maChaine) pour faire défiler la chaîne sur l'afficheur. 
+        Utilisez le retard pour contrôler la vitesse de défilement du texte.
+        Si attendre est faux, le texte défilera pendant que le programme continue. 
+        Si boucle est vrai, le texte se répétera pour toujours. 
+        Si monospace est vrai, les caractères feront toujours 5 Pixels de large.
         """
         pass
 
     def clear(self) -> None:
         """
-        Use clear() to clear micro:bit's display.
+        Utilisez clear () pour effacer l'écran de la micro:bit.
         """
         pass
 
     def get_pixel(self, x: int, y: int) -> int:
         """
-        Use get_pixel(x, y) to return the display's brightness at LED pixel
-        (x,y).
-        Brightness can be from 0 (LED is off) to 9 (maximum LED brightness).
+        get_pixel (x, y) renvoie la luminosité de la LED (x, y). 
+         La luminosité peut être de 0 (LED éteinte) à 9 (luminosité maximale).
         """
         pass
 
-    def set_pixel(self, x: int, y: int, b: int) -> None:
+    def set_pixel(self, x: int, y: int, l: int) -> None:
         """
-        Use set_pixel(x, y, b) to set the display at LED pixel (x,y) to
-        brightness 'b'
-        which can be set between 0 (off) to 9 (full brightness).
+        set_pixel (x, y, b) permet d'allumer la LED de coordonnées (x, y) à la
+        luminosité « l » qui peut être réglé entre 0 (éteinte) à 9 (max).
         """
         pass
 
     def on(self) -> None:
         """
-        Use on() to turn on the display.
+        Utilisez on() pour allumer l'affichage.
         """
         pass
 
     def off(self) -> None:
         """
-        Use off() to turn off the display.
+        Utilisez off() pour éteindre l'affichage.
         """
         pass
 
     def is_on(self) -> bool:
         """
-        Use is_on() to query if the micro:bit's display is on (True) or off
-        (False).
+        Utiliser is_on() pour demander si l'affichage de la micro:bit est allumé (True) 
+        ou éteint (False).
         """
         pass
 
@@ -284,60 +271,61 @@ display = _Display()
 # Pins
 class _Pin(object):
     """
-    A standard pin
+    Une broche standard
     """
     def read_digital(self) -> int:
         """
-        Read the digital value of the pin. 0 for low, 1 for high
+        Lire la valeur numérique de la broche. 0 pour False, 1 pour True.
         """
         pass
 
-    def write_digital(self, value: int) -> None:
+    def write_digital(self, etat: int) -> None:
         """
-        Set the pin to output high if value is 1, or to low, it it is 0.
+        Mets la sortie à l'état haut si la etat est 1, ou à l'état bas si c'est 0.
         """
         pass
 
 
 class _AnaloguePin(_Pin):
     """
-    These pins have ADC & PWM support
+    Ces broches sont pour la CAN et la MLI.
     """
     def read_analog(self) -> int:
         """
-        Read the voltage applied to the pin.
-        Returns the reading as a number between 0 (meaning 0v) and 1023
-        (meaning 3.3v).
+        Lire la tension appliquée à la broche.
+        Renvoie la lecture sous la forme d'un nombre compris entre 0 (c'est-à-dire 0v) et 1023
+        (signifiant 3,3v).
         """
         pass
 
-    def write_analog(self, value: int) -> None:
+    def write_analog(self, RC: int) -> None:
         """
-        Set the pin to output a value between 0 and 1023.
-        """
-        pass
-
-    def set_analog_period(self, period: int) -> None:
-        """
-        Set the period of the PWM signal output to period milliseconds.
+        Émets un signal MLI sur la broche, avec le rapport cyclique proportionnel à la valeur fournie.
+        La valeur peut être un nombre entier ou un nombre à virgule flottante entre 
+        0 (rapport cyclique de 0 %) et 1023 (100%).
         """
         pass
 
-    def set_analog_period_microseconds(self, period: int) -> None:
+    def set_analog_period(self, période: int) -> None:
         """
-        Set the period of the PWM signal output to period microseconds.
+        Définissez la période du signal MLI en millisecondes. La valeur minimale valide est de 1 ms.
+        """
+        pass
+
+    def set_analog_period_microseconds(self, période: int) -> None:
+        """
+        Définissez la période du signal MLI en microsecondes. La valeur minimale valide est de 256 µs.
         """
         pass
 
 
 class _GiantPin(_AnaloguePin):
     """
-    These pins are aranged on the bottom of the board and have holes through
-    them
+    Ces broches sont disposées au bas de la carte et percées de trous.
     """
     def is_touched(self) -> bool:
         """
-        Returns if the pin is being physically touched on the hardware
+        Renvoie 1 si la broche est physiquement touchée sur la carte.
         """
         pass
 
@@ -372,25 +360,23 @@ pin_speaker = _AnaloguePin()
 
 # I2C
 class _I2C(object):
-    def read(self, address: int, n: int, repeat: bool = False) -> bytes:
+    def read(self, addresse: int, n: int, répète: bool = False) -> bytes:
         """
-        Use read(address, n) to read 'n' bytes from the device with the 7-bit
-        address.
-        If repeat is True, no stop bit will be sent.
-        """
-        pass
-
-    def write(self, adress: int, buffer: bytes, repeat: bool = False) -> None:
-        """
-        Use write(address, buffer) to write to the 'buffer' of the device at
-        the 7-bit 'address'.
-        If repeat is True, no stop bit will be sent.
+        Lire 'n' octets du périphérique avec l'adresse sur 7 bits.
+        Si répète est Vrai, aucun bit d'arrêt ne sera envoyé.
         """
         pass
 
-    def init(self, frequency: int, scl: _Pin, sda: _Pin) -> None:
+    def write(self, adresse: int, buffer: bytes, répète: bool = False) -> None:
         """
-        Use init(frequency, scl, sda) to set the bus frequency and pins.
+        Écrire 'n' octets du périphérique avec l'adresse sur 7 bits.
+        Si répète est Vrai, aucun bit d'arrêt ne sera envoyé.
+        """
+        pass
+
+    def init(self, fréquence: int, scl: _Pin, sda: _Pin) -> None:
+        """
+        Définis la fréquence et les broches du bus.
         """
         pass
 
@@ -402,8 +388,8 @@ i2c = _I2C()
 class Image(object):
     def __init__(self, *args: Any) -> None:
         """
-        There are three ways to construct an image
-        First, with a string:
+        Il y a trois façons de construire une image
+        Avec une chaîne :
             Image(
                 '09090:'
                 '99999:'
@@ -411,13 +397,13 @@ class Image(object):
                 '09990:'
                 '00900:'
             )
-            This would create a 5x5 heart shaped image.
-            Numbers go from 0 (off) to 9 (brightest).
-            Note the colon ':' to set the end of a row. You can also use \\n
-        Second with a width/height:
+            Cela créerait une image en forme de cœur 5x5.
+            Les chiffres vont de 0 (éteint) à 9 (le plus lumineux).
+            Notez les deux-points ':' pour définir la fin d'une ligne. 
+        Avec une largeur et une hauteur :
             Image(5, 5)
-            This creates a blank 5x5 image, which you can manipulate
-        Lastly you can pass a buffer with a width and a height
+        Cela crée une image 5x5 vierge, que vous pouvez manipuler.
+        On peut combiner les 2 :
             Image(5, 5, [
                 0, 9, 0, 9, 0,
                 9, 9, 9, 9, 9,
@@ -425,81 +411,77 @@ class Image(object):
                 0, 9, 9, 9, 0,
                 0, 0, 9, 0, 0,
             ])
-            This would create the heart image from the first method.
+            Cela créerait l'image du cœur à partir de la première méthode.
         """
 
     def width(self) -> int:
         """
-        Return the width of the image in pixels.
+        Renvoie la largeur de l'image en pixels.
         """
         pass
 
     def height(self) -> int:
         """
-        Return the height of the image in pixels.
+        Renvoie la hauteur de l'image en pixels.
         """
         pass
 
     def get_pixel(self, x: int, y: int) -> int:
         """
-        Use get_pixel(x, y) to return the image's brightness at LED pixel
-        (x,y).
-        Brightness can be from 0 (LED is off) to 9 (maximum LED brightness).
+        Renvoie la luminosité du point de coordonnées (x, y).
+        La luminosité peut être comprise entre 0 (LED éteinte) et 9 (luminosité maximale).
         """
         pass
 
-    def set_pixel(self, x: int, y: int, b: int) -> None:
+    def set_pixel(self, x: int, y: int, l: int) -> None:
         """
-        Use set_pixel(x, y, b) to set the LED pixel (x,y) in the image to
-        brightness 'b' which can be set between 0 (off) to 9 (full brightness).
+        Défini la luminosité du point de coordonnées (x, y).
+        La luminosité peut être comprise entre 0 (LED éteinte) et 9 (luminosité maximale).
         """
         pass
 
     def shift_left(self, n: int) -> 'Image':
         """
-        Use shift_left(n) to make a copy of the image but moved 'n' pixels to
-        the left.
+        Fait une copie de l'image, mais déplacée de 'n' pixels vers la gauche.
         """
         pass
 
     def shift_right(self, n: int) -> 'Image':
         """
-        Use shift_right(n) to make a copy of the image but moved 'n' pixels to
-        the right.
+        Fait une copie de l'image, mais déplacée de 'n' pixels vers la droite.
         """
         pass
 
     def shift_up(self, n: int) -> 'Image':
         """
-        Use shift_up(n) to make a copy of the image but moved 'n' pixels up.
+        Fait une copie de l'image, mais déplacée de 'n' pixels vers le haut.
         """
         pass
 
     def shift_down(self, n: int) -> 'Image':
         """
-        Use shift_down(n) to make a copy of the image but moved 'n' pixels
-        down.
+        Fait une copie de l'image, mais déplacée de 'n' pixels vers le bas.
         """
         pass
 
     def copy(self) -> 'Image':
         """
-        Use copy() to make a new exact copy of the image.
+        Faire une copie de l'image.
         """
         pass
 
     def crop(self, x1: int, y1: int, x2: int, y2: int) -> 'Image':
         """
-        Use crop(x1, y1, x2, y2) to make a cut-out copy of the image where
-        coordinate (x1,y1) is the top left corner of the cut-out area and
-        coordinate (x2,y2) is the bottom right corner.
+        Faites une copie découpée de l'image où
+        la (x1,y1) est le coin supérieur gauche de la zone de découpe
+        et (x2,y2) est le coin inférieur droit.
         """
         pass
 
     def invert(self) -> 'Image':
         """
-        Use invert() to make a negative copy of the image. Where a pixel was
-        bright or on in the original, it is dim or off in the negative copy.
+        Faire une copie négative de l'image. Un pixel clair ou allumé dans 
+        l'original devient sombre ou éteint dans la copie négative.
         """
         pass
 
@@ -585,49 +567,48 @@ class _UARTSerial(object):
         rx: Optional[_Pin] = None,
     ) -> None:
         """
-        Use init() to set up communication using the default values.
-        Otherwise override the defaults as named arguments.
+        Configure la communication en utilisant les valeurs par défaut.
+        Sinon, remplacez les valeurs par défaut en tant qu'arguments nommés.
         """
         pass
 
     def any(self) -> bool:
         """
-        If there are incoming characters waiting to be read, any() will return
-        True.
-        Otherwise, returns False.
+        S'il y a des caractères entrants en attente de lecture, any() renverra
+        True, sinon, renvoie False.
         """
         pass
 
     def read(self, n: int) -> bytes:
         """
-        Use read() to read characters.
-        Use read(n) to read, at most, 'n' bytes of data.
+        Utilisez read() pour lire les caractères.
+        Utilisez read(n) pour lire au maximum 'n' octets de données.
         """
         pass
 
     def readall(self) -> bytes:
         """
-        Use readall() to read as much data as possible.
+        Lire autant de données que possible.
         """
         pass
 
     def readline(self) -> bytes:
         """
-        Use readline() to read a line that ends with a newline character.
+        Lire une ligne qui se termine par un caractère de saut de ligne.
         """
         pass
 
     def readinto(self, buf: bytes, n: int) -> int:
         """
-        Use readinto(buf) to read bytes into the buffer 'buf'.
-        Use readinto(buff, n) to read, at most, 'n' number of bytes into 'buf'.
+        Utilisez readinto(buf) pour lire des octets dans le tampon 'buf'.
+        Utilisez readinto(buff, n) pour lire au maximum 'n' octets dans 'buf'.
         """
         pass
 
     def write(self, buf: bytes) -> int:
         """
-        Use write(buf) to write the bytes in buffer 'buf' to the connected
-        device.
+        Utilisez write(buf) pour écrire les octets du tampon 'buf' sur le
+        appareil.
         """
         pass
 
@@ -647,30 +628,28 @@ class _SPISerial(object):
         miso: _Pin = pin14,
     ) -> None:
         """
-        Set up communication. Override the defaults for baudrate, mode,
-        SCLK, MOSI and MISO. The default connections are pin13 for SCLK, pin15
-        for MOSI and pin14 for MISO.
+        Initialiser la communication. Remplacer les valeurs par défaut pour le 
+        débit en bauds, le nombre de bit le mode, SCLK, MOSI et MISO. 
         """
         pass
 
     def write(self, buf: bytes) -> None:
         """
-        Use write(buf) to write bytes in buffer 'buf' to the connected device.
+        Utilisez write(buf) pour écrire des octets dans le tampon 'buf' sur le périphérique connecté.
         """
         pass
 
     def read(self, n: int) -> bytes:
         """
-        Use read(n) to read 'n' bytes of data.
+        Utilisez read(n) pour lire 'n' octets de données.
         """
         pass
 
     def write_readinto(self, outbuf: bytes, inbuf: bytes) -> None:
         """
-        Use write_readinto(outbuf, inbuf) to write the 'outbuf' buffer to the
-        connected device and read any response into the 'inbuf' buffer. The
-        length of the buffers should be the same. The buffers can be the same
-        object.
+        Utilisez write_readinto(outbuf, inbuf) pour écrire le tampon 'outbuf' dans le
+        périphérique connecté et lire toute réponse dans le tampon 'inbuf'. La longueur
+        des tampons doit être la même. Les tampons peuvent être les mêmes objets.
         """
         pass
 
@@ -684,14 +663,14 @@ spi = _SPISerial()
 class _Speaker(object):
     def off(self) -> bool:
         """
-        Use off() to turn off the speaker. This does not disable sound output 
-        to an edge connector pin.
+        Éteind le haut-parleur. Cela ne désactive pas la sortie audio
+        sur la broche du connecteur.
         """
         pass
 
     def on(self) -> bool:
         """
-        Use on() to turn on the speaker.
+        Allume le haut-parleur.
         """
         pass
 
@@ -703,35 +682,34 @@ speaker = _Speaker()
 class _Microphone(object):
     def current_event(self) -> str:
         """
-        Returns the name of the last recorded sound event, SoundEvent('loud') 
-        or SoundEvent('quiet').
+        Renvoie le nom du dernier événement sonore enregistré, SoundEvent('loud')
+        ou SoundEvent('quiet').
         """
         pass
 
     def was_event(self, event: str) -> bool:
         """
-        Event is a sound event such as SoundEvent.LOUD or SoundEvent.QUIET. 
-        Returns true if sound was heard at least once since the last call, 
-        otherwise false.
-        The function also clears the sound event history 
-        before returning.
+        Event est un événement sonore tel que SoundEvent.LOUD ou SoundEvent.QUIET.
+        Renvoie vrai si le son a été entendu au moins une fois depuis le dernier appel,
+        sinon faux.
+        La fonction efface également l'historique des événements sonores.
         """
         pass
 
     def is_event(self, event: str) -> bool:
         """
-        Event is a sound event such as SoundEvent.LOUD or SoundEvent.QUIET. 
-        Returns true if sound event is the most recent since the last call, 
-        otherwise false.
-        The function does not clear the sound event history.
+        Event est un événement sonore tel que SoundEvent.LOUD ou SoundEvent.QUIET.
+        Renvoie vrai si l'événement sonore est le plus récent depuis le dernier appel,
+        sinon faux.
+        La fonction n'efface pas l'historique des événements sonores.
         """
         pass
 
     def get_events(self) -> Tuple[str, str]:
         """
-        Returns a tuple of the event history. The most recent is listed 
-        last.
-        The function also clears the sound event history before returning.
+        Renvoie un tuple de l'historique des événements. Le plus récent est listé 
+        en dernier.
+        La fonction efface également l'historique des événements sonores.
         """
         pass
 
@@ -741,17 +719,17 @@ class _Microphone(object):
             value: int,
         ) -> Tuple[str, str]:
         """
-        An event is a sound event, such as SoundEvent.LOUD or SoundEvent.QUIET.
-        Value is the threshold level in the range 0-255. For example, 
-        set_threshold(SoundEvent.LOUD, 250) will only trigger if the sound is 
-        very loud (>= 250).
+        Un événement est un événement sonore, tel que SoundEvent.LOUD ou SoundEvent.QUIET.
+        La valeur est le seuil dans la plage 0-255. Par exemple,
+        set_threshold(SoundEvent.LOUD, 250) ne se déclenchera que si le son est
+        très fort (>= 250).
         """
         pass
 
     def sound_level(self) -> int:
         """
-        Returns a representation of the sound pressure level in the range 0 
-        to 255.
+        Renvoie une représentation du niveau de pression acoustique dans la plage 0
+        à 255.
         """
         pass